CC-1497. Moved sample credentials provider to top-level and referred it in doc.
